refactor: 에너지 저장할 때 시간 float에서 long으로 형변환

+++ Assets/02.Scripts/InGame/Player/Abilities/PlayerHelperInventoryAbility.cs
@@ -123,7 +123,7 @@ public class PlayerHelperInventoryAbility : PlayerAbility, ISaveableAbility
             Grade = (int)_activeHelper.Grade.CurrentGrade,
             Experience = _activeHelper.Experience.CurrentExp,
             Energy = _activeHelper.Energy.Current,
-            EnergySavedAt = UnityEngine.Time.realtimeSinceStartup
+            EnergySavedAt = (long)UnityEngine.Time.realtimeSinceStartup
         };
     }
